13-libraries-refactoring restructure solution
