+++ Common/Common.TasksLibrary/Task1/FileHandler.cs
@@ -1,7 +1,7 @@
-﻿using System.Text;
-using System.Globalization;
+﻿using System.Globalization;
+using System.Text;
 
-namespace Task1Lib
+namespace Common.TasksLibrary.Task1
 {
     public static class FileHandler
     {
