+++ ProgrammingPractice.sln
@@ -3,15 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Task1", "Task1", "{48A69C39-AB5E-4613-9CAB-93E24D1A3516}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tasks", "Tasks", "{48A69C39-AB5E-4613-9CAB-93E24D1A3516}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{AC6FB30A-B748-4093-907A-5C71636E6890}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Task1", "Task1\Task1\Task1.csproj", "{66B8D833-F36E-4FDA-ABE1-ED3453047414}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PracticeTests", "Tests\PracticeTests\PracticeTests.csproj", "{28D48126-9DF6-47B3-8F39-D930DA02953C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Task1Lib", "Task1\Task1Lib\Task1Lib.csproj", "{3F9A2059-6C5A-404F-AE13-92B9BF36AF67}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{E8B9DD88-2BB6-43D2-A27F-5B6F30EB09DD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common.TasksLibrary", "Common\Common.TasksLibrary\Common.TasksLibrary.csproj", "{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tasks.Task1", "Tasks\Tasks.Task1\Tasks.Task1.csproj", "{A4CA7711-A55C-4694-86FF-0F92B9FF58F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,26 +21,26 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{66B8D833-F36E-4FDA-ABE1-ED3453047414}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{66B8D833-F36E-4FDA-ABE1-ED3453047414}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{66B8D833-F36E-4FDA-ABE1-ED3453047414}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{66B8D833-F36E-4FDA-ABE1-ED3453047414}.Release|Any CPU.Build.0 = Release|Any CPU
 		{28D48126-9DF6-47B3-8F39-D930DA02953C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{28D48126-9DF6-47B3-8F39-D930DA02953C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{28D48126-9DF6-47B3-8F39-D930DA02953C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{28D48126-9DF6-47B3-8F39-D930DA02953C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F9A2059-6C5A-404F-AE13-92B9BF36AF67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F9A2059-6C5A-404F-AE13-92B9BF36AF67}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F9A2059-6C5A-404F-AE13-92B9BF36AF67}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F9A2059-6C5A-404F-AE13-92B9BF36AF67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A4CA7711-A55C-4694-86FF-0F92B9FF58F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4CA7711-A55C-4694-86FF-0F92B9FF58F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4CA7711-A55C-4694-86FF-0F92B9FF58F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4CA7711-A55C-4694-86FF-0F92B9FF58F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{66B8D833-F36E-4FDA-ABE1-ED3453047414} = {48A69C39-AB5E-4613-9CAB-93E24D1A3516}
 		{28D48126-9DF6-47B3-8F39-D930DA02953C} = {AC6FB30A-B748-4093-907A-5C71636E6890}
-		{3F9A2059-6C5A-404F-AE13-92B9BF36AF67} = {48A69C39-AB5E-4613-9CAB-93E24D1A3516}
+		{4C45A9CC-D9AC-4750-8CDC-1D5145CB0B0D} = {E8B9DD88-2BB6-43D2-A27F-5B6F30EB09DD}
+		{A4CA7711-A55C-4694-86FF-0F92B9FF58F5} = {48A69C39-AB5E-4613-9CAB-93E24D1A3516}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E45194B0-6EB6-40D4-B102-081853F691AD}
